requirements.txt
Django==4.2.1
djangorestframework==3.15.1
Requests==2.31.0
